remove /synthesis
