+++ modules/gateseq/sequencer_resources.cpp
@@ -5,7 +5,7 @@
  *      Author: willmitchell
  */
 
-#include "sequencers.hpp"
+#include "gateseq.hpp"
 
 /*
  *
