+++ modules/inc/boolean_sequences.hpp
