+++ modules/inc/scales.hpp
