+++ modules/osc3/osc3_ui_implementation.cpp
@@ -6,10 +6,10 @@
  */
 
 
-#include <osc.hpp>
+#include <osc3.hpp>
 #include "user_interface.hpp"
 
-void ViaOsc::ViaOscUI::initialize(void) {
+void ViaOsc3::ViaOsc3UI::initialize(void) {
 
 #ifdef BUILD_VIRTUAL
 
@@ -44,12 +44,12 @@ void ViaOsc::ViaOscUI::initialize(void) {
 	transition(&ViaUI::defaultMenu);
 }
 
-void oscTouchLink (void * uiVoid) {
-	ViaOsc::ViaOscUI * ui = (ViaOsc::ViaOscUI *) uiVoid;
+void osc3TouchLink (void * uiVoid) {
+	ViaOsc3::ViaOsc3UI * ui = (ViaOsc3::ViaOsc3UI *) uiVoid;
 	ui->dispatch(SENSOR_EVENT_SIG);
 }
 
-void ViaOsc::ViaOscUI::recallModuleState(void) {
+void ViaOsc3::ViaOsc3UI::recallModuleState(void) {
 
 	button1Mode %= this_module.numButton1Modes;
 	button2Mode %= this_module.numButton2Modes;
@@ -67,7 +67,7 @@ void ViaOsc::ViaOscUI::recallModuleState(void) {
 
 }
 
-void ViaOsc::ViaOscUI::uiSetLEDs(int mode) {
+void ViaOsc3::ViaOsc3UI::uiSetLEDs(int mode) {
 
 	/// This connects the UI library's LED setting calls to the enumerated LED handler of the this_module.
 	/// This could probably go away with some refactoring.
@@ -76,7 +76,7 @@ void ViaOsc::ViaOscUI::uiSetLEDs(int mode) {
 
 }
 
-void ViaOsc::ViaOscUI::defaultEnterMenuCallback(void) {
+void ViaOsc3::ViaOsc3UI::defaultEnterMenuCallback(void) {
 
 	/// Executed when the module returns to an idle state, waiting for a touch event
 	this_module.runtimeDisplay = 1;
@@ -84,23 +84,23 @@ void ViaOsc::ViaOscUI::defaultEnterMenuCallback(void) {
 	this_module.setLEDD(this_module.chordMode);
 
 }
-void ViaOsc::ViaOscUI::newModeEnterMenuCallback(void) {
+void ViaOsc3::ViaOsc3UI::newModeEnterMenuCallback(void) {
 
-//	transition(&ViaOsc::ViaOscUI::defaultMenu);
-
-}
-void ViaOsc::ViaOscUI::newAuxModeEnterMenuCallback(void) {
-
-	transition(&ViaOsc::ViaOscUI::defaultMenu);
+//	transition(&ViaOsc3::ViaOsc3UI::defaultMenu);
 
 }
-void ViaOsc::ViaOscUI::presetEnterMenuCallback(void) {
+void ViaOsc3::ViaOsc3UI::newAuxModeEnterMenuCallback(void) {
 
-	transition(&ViaOsc::ViaOscUI::defaultMenu);
+	transition(&ViaOsc3::ViaOsc3UI::defaultMenu);
+
+}
+void ViaOsc3::ViaOsc3UI::presetEnterMenuCallback(void) {
+
+	transition(&ViaOsc3::ViaOsc3UI::defaultMenu);
 
 }
 
-void ViaOsc::ViaOscUI::button1EnterMenuCallback(void) {
+void ViaOsc3::ViaOsc3UI::button1EnterMenuCallback(void) {
 
 	this_module.runtimeDisplay = 0;
 	this_module.clearLEDs();
@@ -109,7 +109,7 @@ void ViaOsc::ViaOscUI::button1EnterMenuCallback(void) {
 	resetTimerMenu();
 
 }
-void ViaOsc::ViaOscUI::button2EnterMenuCallback(void) {
+void ViaOsc3::ViaOsc3UI::button2EnterMenuCallback(void) {
 
 	this_module.runtimeDisplay = 0;
 	this_module.clearLEDs();
@@ -118,7 +118,7 @@ void ViaOsc::ViaOscUI::button2EnterMenuCallback(void) {
 	resetTimerMenu();
 
 }
-void ViaOsc::ViaOscUI::button3EnterMenuCallback(void) {
+void ViaOsc3::ViaOsc3UI::button3EnterMenuCallback(void) {
 
 	this_module.runtimeDisplay = 0;
 	this_module.clearLEDs();
@@ -127,7 +127,7 @@ void ViaOsc::ViaOscUI::button3EnterMenuCallback(void) {
 	resetTimerMenu();
 
 }
-void ViaOsc::ViaOscUI::button4EnterMenuCallback(void) {
+void ViaOsc3::ViaOsc3UI::button4EnterMenuCallback(void) {
 
 	this_module.runtimeDisplay = 0;
 	this_module.clearLEDs();
@@ -136,7 +136,7 @@ void ViaOsc::ViaOscUI::button4EnterMenuCallback(void) {
 	resetTimerMenu();
 
 }
-void ViaOsc::ViaOscUI::button5EnterMenuCallback(void) {
+void ViaOsc3::ViaOsc3UI::button5EnterMenuCallback(void) {
 
 	this_module.runtimeDisplay = 0;
 	this_module.clearLEDs();
@@ -146,7 +146,7 @@ void ViaOsc::ViaOscUI::button5EnterMenuCallback(void) {
 	resetTimerMenu();
 
 }
-void ViaOsc::ViaOscUI::button6EnterMenuCallback(void) {
+void ViaOsc3::ViaOsc3UI::button6EnterMenuCallback(void) {
 
 	this_module.runtimeDisplay = 0;
 	this_module.clearLEDs();
@@ -154,14 +154,14 @@ void ViaOsc::ViaOscUI::button6EnterMenuCallback(void) {
 	this_module.setLEDs(button6Mode);
 	resetTimerMenu();
 }
-void ViaOsc::ViaOscUI::aux1EnterMenuCallback(void) {
+void ViaOsc3::ViaOsc3UI::aux1EnterMenuCallback(void) {
 
 	/// Executed after a press is detected at button 1 from the shift state
 	this_module.runtimeDisplay = 0;
 	resetTimerMenu();
 
 }
-void ViaOsc::ViaOscUI::aux2EnterMenuCallback(void) {
+void ViaOsc3::ViaOsc3UI::aux2EnterMenuCallback(void) {
 
 	/// Executed after a press is detected at button 3 from the shift state
 	this_module.runtimeDisplay = 0;
@@ -169,7 +169,7 @@ void ViaOsc::ViaOscUI::aux2EnterMenuCallback(void) {
 
 }
 
-void ViaOsc::ViaOscUI::aux2AltEnterMenuCallback(void) {
+void ViaOsc3::ViaOsc3UI::aux2AltEnterMenuCallback(void) {
 
 	/// Executed after a press is detected at button 2 from the shift state
 	this_module.runtimeDisplay = 0;
@@ -177,14 +177,14 @@ void ViaOsc::ViaOscUI::aux2AltEnterMenuCallback(void) {
 
 }
 
-void ViaOsc::ViaOscUI::aux3EnterMenuCallback(void) {
+void ViaOsc3::ViaOsc3UI::aux3EnterMenuCallback(void) {
 
 	/// Executed after a press is detected at button 4 from the shift state
 	this_module.runtimeDisplay = 0;
 	resetTimerMenu();
 
 }
-void ViaOsc::ViaOscUI::aux4EnterMenuCallback(void) {
+void ViaOsc3::ViaOsc3UI::aux4EnterMenuCallback(void) {
 
 	/// Executed after a press is detected at button 6 from the shift state
 	this_module.runtimeDisplay = 0;
@@ -195,155 +195,155 @@ void ViaOsc::ViaOscUI::aux4EnterMenuCallback(void) {
 
 /// Executed after exiting a button menu with a release event in less than half a second (roughly)
 
-void ViaOsc::ViaOscUI::button1TapCallback(void) {
+void ViaOsc3::ViaOsc3UI::button1TapCallback(void) {
 
 	button1Mode = button1Mode > 4 ? 5 : button1Mode + 1;
 	storeMode(button1Mode, BUTTON1_MASK, BUTTON1_SHIFT);
 	this_module.handleButton1ModeChange(button1Mode);
 	this_module.clearLEDs();
 	this_module.setLEDs(button1Mode);
-	transition(&ViaOsc::ViaOscUI::newModeMenu);
+	transition(&ViaOsc3::ViaOsc3UI::newModeMenu);
 
 }
-void ViaOsc::ViaOscUI::button2TapCallback(void) {
+void ViaOsc3::ViaOsc3UI::button2TapCallback(void) {
 
 	button2Mode = incrementModeAndStore(button2Mode, BUTTON2_MASK, this_module.numButton2Modes, BUTTON2_SHIFT);
 	this_module.handleButton2ModeChange(button2Mode);
 	this_module.clearLEDs();
 	this_module.setLEDs(button2Mode);
-	transition(&ViaOsc::ViaOscUI::newModeMenu);
+	transition(&ViaOsc3::ViaOsc3UI::newModeMenu);
 
 }
-void ViaOsc::ViaOscUI::button3TapCallback(void) {
+void ViaOsc3::ViaOsc3UI::button3TapCallback(void) {
 
 	button3Mode = incrementModeAndStore(button3Mode, BUTTON3_MASK, this_module.numButton3Modes, BUTTON3_SHIFT);
 	this_module.handleButton3ModeChange(button3Mode);
 	this_module.clearLEDs();
 	this_module.setLEDs(button3Mode);
-	transition(&ViaOsc::ViaOscUI::newModeMenu);
+	transition(&ViaOsc3::ViaOsc3UI::newModeMenu);
 
 }
-void ViaOsc::ViaOscUI::button4TapCallback(void) {
+void ViaOsc3::ViaOsc3UI::button4TapCallback(void) {
 
 	button1Mode = button1Mode <= 0 ? 0 : button1Mode - 1;
 	storeMode(button1Mode, BUTTON1_MASK, BUTTON1_SHIFT);
 	this_module.handleButton1ModeChange(button1Mode);
 	this_module.clearLEDs();
 	this_module.setLEDs(button1Mode);
-	transition(&ViaOsc::ViaOscUI::newModeMenu);
+	transition(&ViaOsc3::ViaOsc3UI::newModeMenu);
 
 }
-void ViaOsc::ViaOscUI::button5TapCallback(void) {
+void ViaOsc3::ViaOsc3UI::button5TapCallback(void) {
 
 	button5Mode = incrementModeAndStore(button5Mode, BUTTON5_MASK, this_module.numButton5Modes, BUTTON5_SHIFT);
 	this_module.handleButton5ModeChange(button5Mode);
 	this_module.clearLEDs();
 	this_module.setLEDs(button5Mode);
 	this_module.showQuant();
-	transition(&ViaOsc::ViaOscUI::newModeMenu);
+	transition(&ViaOsc3::ViaOsc3UI::newModeMenu);
 
 }
-void ViaOsc::ViaOscUI::button6TapCallback(void) {
+void ViaOsc3::ViaOsc3UI::button6TapCallback(void) {
 
 	button6Mode = incrementModeAndStore(button6Mode, BUTTON6_MASK, this_module.numButton6Modes, BUTTON6_SHIFT);
 	this_module.handleButton6ModeChange(button6Mode);
 	this_module.clearLEDs();
 	this_module.setLEDs(button6Mode);
-	transition(&ViaOsc::ViaOscUI::newModeMenu);
+	transition(&ViaOsc3::ViaOsc3UI::newModeMenu);
 
 }
 
-void ViaOsc::ViaOscUI::aux1TapCallback(void) {
+void ViaOsc3::ViaOsc3UI::aux1TapCallback(void) {
 
-	transition(&ViaOsc::ViaOscUI::newModeMenu);
-
-}
-
-void ViaOsc::ViaOscUI::aux2TapCallback(void) {
-
-	transition(&ViaOsc::ViaOscUI::newModeMenu);
+	transition(&ViaOsc3::ViaOsc3UI::newModeMenu);
 
 }
 
-void ViaOsc::ViaOscUI::aux2AltTapCallback(void) {
+void ViaOsc3::ViaOsc3UI::aux2TapCallback(void) {
 
-	transition(&ViaOsc::ViaOscUI::newModeMenu);
-
-}
-
-void ViaOsc::ViaOscUI::aux3TapCallback(void) {
-
-	transition(&ViaOsc::ViaOscUI::newModeMenu);
+	transition(&ViaOsc3::ViaOsc3UI::newModeMenu);
 
 }
 
-void ViaOsc::ViaOscUI::aux4TapCallback(void) {
+void ViaOsc3::ViaOsc3UI::aux2AltTapCallback(void) {
 
-	transition(&ViaOsc::ViaOscUI::newModeMenu);
+	transition(&ViaOsc3::ViaOsc3UI::newModeMenu);
+
+}
+
+void ViaOsc3::ViaOsc3UI::aux3TapCallback(void) {
+
+	transition(&ViaOsc3::ViaOsc3UI::newModeMenu);
+
+}
+
+void ViaOsc3::ViaOsc3UI::aux4TapCallback(void) {
+
+	transition(&ViaOsc3::ViaOsc3UI::newModeMenu);
 
 }
 
 
 /// Executed after exiting a button menu with a release event in more than half a second (roughly)
 
-void ViaOsc::ViaOscUI::button1HoldCallback(void) {
+void ViaOsc3::ViaOsc3UI::button1HoldCallback(void) {
 
-	transition(&ViaOsc::ViaOscUI::defaultMenu);
-
-}
-void ViaOsc::ViaOscUI::button2HoldCallback(void) {
-
-	transition(&ViaOsc::ViaOscUI::defaultMenu);
+	transition(&ViaOsc3::ViaOsc3UI::defaultMenu);
 
 }
-void ViaOsc::ViaOscUI::button3HoldCallback(void) {
+void ViaOsc3::ViaOsc3UI::button2HoldCallback(void) {
 
-	transition(&ViaOsc::ViaOscUI::defaultMenu);
-
-}
-void ViaOsc::ViaOscUI::button4HoldCallback(void) {
-
-	transition(&ViaOsc::ViaOscUI::defaultMenu);
+	transition(&ViaOsc3::ViaOsc3UI::defaultMenu);
 
 }
-void ViaOsc::ViaOscUI::button5HoldCallback(void) {
+void ViaOsc3::ViaOsc3UI::button3HoldCallback(void) {
 
-	transition(&ViaOsc::ViaOscUI::defaultMenu);
-
-}
-void ViaOsc::ViaOscUI::button6HoldCallback(void) {
-
-	transition(&ViaOsc::ViaOscUI::defaultMenu);
+	transition(&ViaOsc3::ViaOsc3UI::defaultMenu);
 
 }
+void ViaOsc3::ViaOsc3UI::button4HoldCallback(void) {
 
-void ViaOsc::ViaOscUI::aux1HoldCallback(void) {
+	transition(&ViaOsc3::ViaOsc3UI::defaultMenu);
 
-	transition(&ViaOsc::ViaOscUI::button5Menu);
+}
+void ViaOsc3::ViaOsc3UI::button5HoldCallback(void) {
+
+	transition(&ViaOsc3::ViaOsc3UI::defaultMenu);
+
+}
+void ViaOsc3::ViaOsc3UI::button6HoldCallback(void) {
+
+	transition(&ViaOsc3::ViaOsc3UI::defaultMenu);
 
 }
 
-void ViaOsc::ViaOscUI::aux2HoldCallback(void) {
+void ViaOsc3::ViaOsc3UI::aux1HoldCallback(void) {
 
-	transition(&ViaOsc::ViaOscUI::button5Menu);
-
-}
-
-void ViaOsc::ViaOscUI::aux2AltHoldCallback(void) {
-
-	transition(&ViaOsc::ViaOscUI::button5Menu);
+	transition(&ViaOsc3::ViaOsc3UI::button5Menu);
 
 }
 
-void ViaOsc::ViaOscUI::aux3HoldCallback(void) {
+void ViaOsc3::ViaOsc3UI::aux2HoldCallback(void) {
 
-	transition(&ViaOsc::ViaOscUI::button5Menu);
+	transition(&ViaOsc3::ViaOsc3UI::button5Menu);
 
 }
 
-void ViaOsc::ViaOscUI::aux4HoldCallback(void) {
+void ViaOsc3::ViaOsc3UI::aux2AltHoldCallback(void) {
 
-	transition(&ViaOsc::ViaOscUI::button5Menu);
+	transition(&ViaOsc3::ViaOsc3UI::button5Menu);
+
+}
+
+void ViaOsc3::ViaOsc3UI::aux3HoldCallback(void) {
+
+	transition(&ViaOsc3::ViaOsc3UI::button5Menu);
+
+}
+
+void ViaOsc3::ViaOsc3UI::aux4HoldCallback(void) {
+
+	transition(&ViaOsc3::ViaOsc3UI::button5Menu);
 
 }
