+++ modules/scanner/three_axis_scanner.cpp
@@ -5,7 +5,7 @@
  *      Author: willmitchell
  */
 
-#include "signal_processors.hpp"
+#include "scanner.hpp"
 
 
 void ThreeAxisScanner::parseControls(ViaControls * controls) {
