+++ modules/sync/pll_controller.cpp
@@ -5,7 +5,7 @@
  *      Author: willmitchell
  */
 
-#include <oscillators.hpp>
+#include <sync.hpp>
 
 
 /*
