--- synthesis/inc/oscillators.hpp
@@ -1,534 +1,0 @@
-/*
- * oscillators.h
- *
- *  Created on: Aug 18, 2018
- *      Author: willmitchell
- */
-
-#ifndef INC_OSCILLATORS_HPP_
-#define INC_OSCILLATORS_HPP_
-
-#include <via_platform_binding.hpp>
-#include "tables.hpp"
-#include "scales.hpp"
-
-#define WAVETABLE_LENGTH 33554432
-#define NEGATIVE_WAVETABLE_LENGTH -33554432 // wavetable length in 16 bit fixed point (512 << 16)
-#define AT_B_PHASE 16777216 // wavetable midpoint in 16 bit fixed point (256 << 16)
-
-// phase events
-#define NO_EVENT 0
-#define AT_A_FROM_RELEASE -WAVETABLE_LENGTH + 1
-#define AT_A_FROM_ATTACK WAVETABLE_LENGTH - 1
-#define AT_B_FROM_ATTACK -1
-#define AT_B_FROM_RELEASE 1
-
-#define phaseModPWM_0 {0, 1310719, 2621439, 3932159, 5242879, 6553599, 7864319, 9175039, 10485759, 11796479, 13107199, 14417919, 15728639, 16842751, 17170431, 17498111, 17825791, 18153471, 18481151, 18808831, 19136511, 19464191, 19791871, 20119551, 20447231, 20774911, 21102591, 21430271, 21757951, 22085631, 22413311, 22740991, 23068671, 23396351, 23724031, 24051711, 24379391, 24707071, 25034751, 25362431, 25690111, 26017791, 26345471, 26673151, 27000831, 27328511, 27656191, 27983871, 28311551, 28639231, 28966911, 29294591, 29622271, 29949951, 30277631, 30605311, 30932991, 31260671, 31588351, 31916031, 32243711, 32571391, 32899071, 33226751, 33554431}
-
-#define phaseModPWM_1 {0, 1198372, 2396745, 3595117, 4793490, 5991862, 7190235, 8388607, 9586980, 10785352, 11983725, 13182097, 14380470, 15578842, 16777215, 17112759, 17448304, 17783848, 18119392, 18454937, 18790481, 19126025, 19461569, 19797114, 20132658, 20468202, 20803747, 21139291, 21474835, 21810380, 22145924, 22481468, 22817013, 23152557, 23488101, 23823646, 24159190, 24494734, 24830278, 25165823, 25501367, 25836911, 26172456, 26508000, 26843544, 27179089, 27514633, 27850177, 28185722, 28521266, 28856810, 29192354, 29527899, 29863443, 30198987, 30534532, 30870076, 31205620, 31541165, 31876709, 32212253, 32547798, 32883342, 33218886, 33554431}
-
-#define phaseModPWM_2 {0, 1103764, 2207528, 3311292, 4415056, 5518820, 6622585, 7726349, 8830113, 9933877, 11037641, 12141405, 13245170, 14348934, 15452698, 16556462, 17052251, 17396047, 17739842, 18083638, 18427433, 18771228, 19115024, 19458819, 19802615, 20146410, 20490205, 20834001, 21177796, 21521592, 21865387, 22209182, 22552978, 22896773, 23240569, 23584364, 23928159, 24271955, 24615750, 24959546, 25303341, 25647136, 25990932, 26334727, 26678523, 27022318, 27366113, 27709909, 28053704, 28397500, 28741295, 29085090, 29428886, 29772681, 30116477, 30460272, 30804067, 31147863, 31491658, 31835454, 32179249, 32523044, 32866840, 33210635, 33554431}
-
-#define phaseModPWM_3 {0, 1023000, 2046001, 3069002, 4092003, 5115004, 6138005, 7161006, 8184007, 9207008, 10230009, 11253010, 12276011, 13299012, 14322013, 15345014, 16368015, 16988693, 17341155, 17693618, 18046080, 18398543, 18751005, 19103468, 19455930, 19808393, 20160855, 20513318, 20865780, 21218243, 21570705, 21923168, 22275630, 22628093, 22980555, 23333018, 23685480, 24037943, 24390405, 24742868, 25095330, 25447793, 25800255, 26152718, 26505180, 26857643, 27210105, 27562568, 27915030, 28267493, 28619955, 28972418, 29324880, 29677343, 30029805, 30382268, 30734730, 31087193, 31439655, 31792118, 32144580, 32497043, 32849505, 33201968, 33554431}
-
-#define phaseModPWM_4 {0, 953250, 1906501, 2859752, 3813003, 4766254, 5719505, 6672756, 7626007, 8579257, 9532508, 10485759, 11439010, 12392261, 13345512, 14298763, 15252014, 16205264, 16921846, 17283424, 17645002, 18006580, 18368158, 18729736, 19091314, 19452892, 19814470, 20176047, 20537625, 20899203, 21260781, 21622359, 21983937, 22345515, 22707093, 23068671, 23430249, 23791827, 24153405, 24514982, 24876560, 25238138, 25599716, 25961294, 26322872, 26684450, 27046028, 27407606, 27769184, 28130762, 28492340, 28853918, 29215495, 29577073, 29938651, 30300229, 30661807, 31023385, 31384963, 31746541, 32108119, 32469697, 32831275, 33192853, 33554431}
-
-#define phaseModPWM_5 {0, 892405, 1784810, 2677215, 3569620, 4462025, 5354430, 6246835, 7139240, 8031645, 8924050, 9816455, 10708860, 11601266, 12493671, 13386076, 14278481, 15170886, 16063291, 16851450, 17222628, 17593805, 17964982, 18336160, 18707337, 19078514, 19449692, 19820869, 20192046, 20563224, 20934401, 21305578, 21676756, 22047933, 22419110, 22790288, 23161465, 23532642, 23903820, 24274997, 24646174, 25017352, 25388529, 25759706, 26130884, 26502061, 26873238, 27244416, 27615593, 27986770, 28357948, 28729125, 29100302, 29471480, 29842657, 30213834, 30585012, 30956189, 31327366, 31698544, 32069721, 32440898, 32812076, 33183253, 33554431}
-
-#define phaseModPWM_6 {0, 838860, 1677721, 2516582, 3355443, 4194303, 5033164, 5872025, 6710886, 7549746, 8388607, 9227468, 10066329, 10905190, 11744050, 12582911, 13421772, 14260633, 15099493, 15938354, 16777215, 17158515, 17539816, 17921116, 18302416, 18683717, 19065017, 19446317, 19827618, 20208918, 20590219, 20971519, 21352819, 21734120, 22115420, 22496720, 22878021, 23259321, 23640621, 24021922, 24403222, 24784522, 25165823, 25547123, 25928423, 26309724, 26691024, 27072325, 27453625, 27834925, 28216226, 28597526, 28978826, 29360127, 29741427, 30122727, 30504028, 30885328, 31266628, 31647929, 32029229, 32410529, 32791830, 33173130, 33554431}
-
-#define phaseModPWM_7 {0, 791378, 1582756, 2374134, 3165512, 3956890, 4748268, 5539646, 6331024, 7122402, 7913780, 8705158, 9496537, 10287915, 11079293, 11870671, 12662049, 13453427, 14244805, 15036183, 15827561, 16618939, 17090808, 17482799, 17874790, 18266781, 18658772, 19050763, 19442754, 19834745, 20226736, 20618727, 21010718, 21402709, 21794700, 22186691, 22578682, 22970673, 23362664, 23754655, 24146646, 24538637, 24930628, 25322619, 25714610, 26106601, 26498592, 26890583, 27282574, 27674565, 28066556, 28458547, 28850538, 29242529, 29634520, 30026511, 30418502, 30810493, 31202484, 31594475, 31986466, 32378457, 32770448, 33162439, 33554431}
-
-#define phaseModPWM_8 {0, 748982, 1497965, 2246948, 2995931, 3744914, 4493897, 5242879, 5991862, 6740845, 7489828, 8238811, 8987794, 9736776, 10485759, 11234742, 11983725, 12732708, 13481691, 14230673, 14979656, 15728639, 16477622, 17019194, 17422493, 17825791, 18229089, 18632388, 19035686, 19438985, 19842283, 20245582, 20648880, 21052179, 21455477, 21858775, 22262074, 22665372, 23068671, 23471969, 23875268, 24278566, 24681865, 25085163, 25488462, 25891760, 26295058, 26698357, 27101655, 27504954, 27908252, 28311551, 28714849, 29118148, 29521446, 29924744, 30328043, 30731341, 31134640, 31537938, 31941237, 32344535, 32747834, 33151132, 33554431}
-
-#define phaseModPWM_9 {0, 710898, 1421797, 2132696, 2843595, 3554494, 4265393, 4976292, 5687191, 6398090, 7108989, 7819888, 8530787, 9241686, 9952585, 10663484, 11374383, 12085282, 12796181, 13507080, 14217979, 14928878, 15639777, 16350676, 16943326, 17358604, 17773881, 18189159, 18604436, 19019714, 19434992, 19850269, 20265547, 20680825, 21096102, 21511380, 21926657, 22341935, 22757213, 23172490, 23587768, 24003045, 24418323, 24833601, 25248878, 25664156, 26079433, 26494711, 26909989, 27325266, 27740544, 28155822, 28571099, 28986377, 29401654, 29816932, 30232210, 30647487, 31062765, 31478042, 31893320, 32308598, 32723875, 33139153, 33554431}
-
-#define phaseModPWM_10 {0, 676500, 1353001, 2029501, 2706002, 3382503, 4059003, 4735504, 5412004, 6088505, 6765006, 7441506, 8118007, 8794508, 9471008, 10147509, 10824009, 11500510, 12177011, 12853511, 13530012, 14206513, 14883013, 15559514, 16236014, 16862813, 17290803, 17718793, 18146784, 18574774, 19002764, 19430754, 19858744, 20286735, 20714725, 21142715, 21570705, 21998695, 22426686, 22854676, 23282666, 23710656, 24138646, 24566636, 24994627, 25422617, 25850607, 26278597, 26706587, 27134578, 27562568, 27990558, 28418548, 28846538, 29274529, 29702519, 30130509, 30558499, 30986489, 31414480, 31842470, 32270460, 32698450, 33126440, 33554431}
-
-#define phaseModPWM_11 {0, 645277, 1290555, 1935832, 2581110, 3226387, 3871665, 4516942, 5162220, 5807497, 6452775, 7098052, 7743330, 8388607, 9033885, 9679162, 10324440, 10969717, 11614995, 12260272, 12905550, 13550827, 14196105, 14841382, 15486660, 16131937, 16777215, 17218721, 17660226, 18101732, 18543238, 18984743, 19426249, 19867755, 20309260, 20750766, 21192272, 21633777, 22075283, 22516789, 22958294, 23399800, 23841306, 24282811, 24724317, 25165823, 25607328, 26048834, 26490340, 26931845, 27373351, 27814857, 28256362, 28697868, 29139374, 29580879, 30022385, 30463891, 30905396, 31346902, 31788408, 32229913, 32671419, 33112925, 33554431}
-
-#define phaseModPWM_12 {0, 616809, 1233618, 1850428, 2467237, 3084046, 3700856, 4317665, 4934475, 5551284, 6168093, 6784903, 7401712, 8018522, 8635331, 9252140, 9868950, 10485759, 11102569, 11719378, 12336187, 12952997, 13569806, 14186616, 14803425, 15420234, 16037044, 16653853, 17141937, 17597840, 18053742, 18509645, 18965547, 19421450, 19877353, 20333255, 20789158, 21245060, 21700963, 22156866, 22612768, 23068671, 23524573, 23980476, 24436379, 24892281, 25348184, 25804086, 26259989, 26715892, 27171794, 27627697, 28083599, 28539502, 28995405, 29451307, 29907210, 30363112, 30819015, 31274918, 31730820, 32186723, 32642625, 33098528, 33554431}
-
-#define phaseModPWM_13 {0, 590747, 1181494, 1772241, 2362988, 2953735, 3544482, 4135229, 4725976, 5316723, 5907470, 6498217, 7088964, 7679711, 8270458, 8861205, 9451952, 10042699, 10633446, 11224193, 11814940, 12405687, 12996434, 13587181, 14177928, 14768675, 15359422, 15950169, 16540916, 17059977, 17531247, 18002517, 18473787, 18945057, 19416328, 19887598, 20358868, 20830138, 21301408, 21772678, 22243948, 22715218, 23186488, 23657758, 24129029, 24600299, 25071569, 25542839, 26014109, 26485379, 26956649, 27427919, 27899189, 28370459, 28841730, 29313000, 29784270, 30255540, 30726810, 31198080, 31669350, 32140620, 32611890, 33083160, 33554431}
-
-#define phaseModPWM_14 {0, 566797, 1133595, 1700393, 2267191, 2833989, 3400786, 3967584, 4534382, 5101180, 5667978, 6234776, 6801573, 7368371, 7935169, 8501967, 9068765, 9635562, 10202360, 10769158, 11335956, 11902754, 12469552, 13036349, 13603147, 14169945, 14736743, 15303541, 15870338, 16437136, 16972299, 17460009, 17947718, 18435428, 18923138, 19410848, 19898557, 20386267, 20873977, 21361687, 21849396, 22337106, 22824816, 23312526, 23800235, 24287945, 24775655, 25263365, 25751074, 26238784, 26726494, 27214204, 27701913, 28189623, 28677333, 29165043, 29652752, 30140462, 30628172, 31115882, 31603591, 32091301, 32579011, 33066721, 33554431}
-
-#define phaseModPWM_15 {0, 544714, 1089429, 1634144, 2178859, 2723573, 3268288, 3813003, 4357718, 4902433, 5447147, 5991862, 6536577, 7081292, 7626007, 8170721, 8715436, 9260151, 9804866, 10349580, 10894295, 11439010, 11983725, 12528440, 13073154, 13617869, 14162584, 14707299, 15252014, 15796728, 16341443, 16878283, 17383620, 17888958, 18394296, 18899634, 19404972, 19910309, 20415647, 20920985, 21426323, 21931661, 22436999, 22942336, 23447674, 23953012, 24458350, 24963688, 25469025, 25974363, 26479701, 26985039, 27490377, 27995715, 28501052, 29006390, 29511728, 30017066, 30522404, 31027741, 31533079, 32038417, 32543755, 33049093, 33554431}
-
-#define phaseModPWM_16 {0, 524287, 1048575, 1572863, 2097151, 2621439, 3145727, 3670015, 4194303, 4718591, 5242879, 5767167, 6291455, 6815743, 7340031, 7864319, 8388607, 8912895, 9437183, 9961471, 10485759, 11010047, 11534335, 12058623, 12582911, 13107199, 13631487, 14155775, 14680063, 15204351, 15728639, 16252927, 16777215, 17301503, 17825791, 18350079, 18874367, 19398655, 19922943, 20447231, 20971519, 21495807, 22020095, 22544383, 23068671, 23592959, 24117247, 24641535, 25165823, 25690111, 26214399, 26738687, 27262975, 27787263, 28311551, 28835839, 29360127, 29884415, 30408703, 30932991, 31457279, 31981567, 32505855, 33030143, 33554431}
-
-#define phaseModPWM_17 {0, 505337, 1010675, 1516013, 2021351, 2526689, 3032026, 3537364, 4042702, 4548040, 5053378, 5558715, 6064053, 6569391, 7074729, 7580067, 8085405, 8590742, 9096080, 9601418, 10106756, 10612094, 11117431, 11622769, 12128107, 12633445, 13138783, 13644121, 14149458, 14654796, 15160134, 15665472, 16170810, 16676147, 17212987, 17757702, 18302416, 18847131, 19391846, 19936561, 20481276, 21025990, 21570705, 22115420, 22660135, 23204850, 23749564, 24294279, 24838994, 25383709, 25928423, 26473138, 27017853, 27562568, 28107283, 28651997, 29196712, 29741427, 30286142, 30830857, 31375571, 31920286, 32465001, 33009716, 33554431}
-
-#define phaseModPWM_18 {0, 487709, 975419, 1463129, 1950839, 2438548, 2926258, 3413968, 3901678, 4389387, 4877097, 5364807, 5852517, 6340226, 6827936, 7315646, 7803356, 8291065, 8778775, 9266485, 9754195, 10241904, 10729614, 11217324, 11705034, 12192743, 12680453, 13168163, 13655873, 14143582, 14631292, 15119002, 15606712, 16094421, 16582131, 17117294, 17684092, 18250889, 18817687, 19384485, 19951283, 20518081, 21084878, 21651676, 22218474, 22785272, 23352070, 23918868, 24485665, 25052463, 25619261, 26186059, 26752857, 27319654, 27886452, 28453250, 29020048, 29586846, 30153644, 30720441, 31287239, 31854037, 32420835, 32987633, 33554431}
-
-#define phaseModPWM_19 {0, 471270, 942540, 1413810, 1885080, 2356350, 2827620, 3298890, 3770160, 4241430, 4712700, 5183971, 5655241, 6126511, 6597781, 7069051, 7540321, 8011591, 8482861, 8954131, 9425401, 9896672, 10367942, 10839212, 11310482, 11781752, 12253022, 12724292, 13195562, 13666832, 14138102, 14609373, 15080643, 15551913, 16023183, 16494453, 17013514, 17604261, 18195008, 18785755, 19376502, 19967249, 20557996, 21148743, 21739490, 22330237, 22920984, 23511731, 24102478, 24693225, 25283972, 25874719, 26465466, 27056213, 27646960, 28237707, 28828454, 29419201, 30009948, 30600695, 31191442, 31782189, 32372936, 32963683, 33554431}
-
-#define phaseModPWM_20 {0, 455902, 911805, 1367707, 1823610, 2279512, 2735415, 3191318, 3647220, 4103123, 4559025, 5014928, 5470831, 5926733, 6382636, 6838538, 7294441, 7750344, 8206246, 8662149, 9118051, 9573954, 10029857, 10485759, 10941662, 11397564, 11853467, 12309370, 12765272, 13221175, 13677077, 14132980, 14588883, 15044785, 15500688, 15956590, 16412493, 16900577, 17517386, 18134196, 18751005, 19367814, 19984624, 20601433, 21218243, 21835052, 22451861, 23068671, 23685480, 24302290, 24919099, 25535908, 26152718, 26769527, 27386337, 28003146, 28619955, 29236765, 29853574, 30470384, 31087193, 31704002, 32320812, 32937621, 33554431}
-
-#define phaseModPWM_21 {0, 441505, 883011, 1324517, 1766022, 2207528, 2649034, 3090539, 3532045, 3973551, 4415056, 4856562, 5298068, 5739573, 6181079, 6622585, 7064090, 7505596, 7947102, 8388607, 8830113, 9271619, 9713124, 10154630, 10596136, 11037641, 11479147, 11920653, 12362158, 12803664, 13245170, 13686675, 14128181, 14569687, 15011192, 15452698, 15894204, 16335709, 16777215, 17422493, 18067770, 18713048, 19358325, 20003603, 20648880, 21294158, 21939435, 22584713, 23229990, 23875268, 24520545, 25165823, 25811100, 26456378, 27101655, 27746933, 28392210, 29037488, 29682765, 30328043, 30973320, 31618598, 32263875, 32909153, 33554431}
-
-#define phaseModPWM_22 {0, 427990, 855980, 1283970, 1711960, 2139950, 2567941, 2995931, 3423921, 3851911, 4279901, 4707892, 5135882, 5563872, 5991862, 6419852, 6847843, 7275833, 7703823, 8131813, 8559803, 8987794, 9415784, 9843774, 10271764, 10699754, 11127744, 11555735, 11983725, 12411715, 12839705, 13267695, 13695686, 14123676, 14551666, 14979656, 15407646, 15835637, 16263627, 16691617, 17318416, 17994916, 18671417, 19347917, 20024418, 20700919, 21377419, 22053920, 22730421, 23406921, 24083422, 24759922, 25436423, 26112924, 26789424, 27465925, 28142426, 28818926, 29495427, 30171927, 30848428, 31524929, 32201429, 32877930, 33554431}
-
-#define phaseModPWM_23 {0, 415277, 830555, 1245832, 1661110, 2076388, 2491665, 2906943, 3322220, 3737498, 4152776, 4568053, 4983331, 5398608, 5813886, 6229164, 6644441, 7059719, 7474997, 7890274, 8305552, 8720829, 9136107, 9551385, 9966662, 10381940, 10797217, 11212495, 11627773, 12043050, 12458328, 12873605, 13288883, 13704161, 14119438, 14534716, 14949994, 15365271, 15780549, 16195826, 16611104, 17203754, 17914653, 18625552, 19336451, 20047350, 20758249, 21469148, 22180047, 22890946, 23601845, 24312744, 25023643, 25734542, 26445441, 27156340, 27867239, 28578138, 29289037, 29999936, 30710835, 31421734, 32132633, 32843532, 33554431}
-
-#define phaseModPWM_24 {0, 403298, 806596, 1209895, 1613193, 2016492, 2419790, 2823089, 3226387, 3629686, 4032984, 4436282, 4839581, 5242879, 5646178, 6049476, 6452775, 6856073, 7259372, 7662670, 8065968, 8469267, 8872565, 9275864, 9679162, 10082461, 10485759, 10889058, 11292356, 11695655, 12098953, 12502251, 12905550, 13308848, 13712147, 14115445, 14518744, 14922042, 15325341, 15728639, 16131937, 16535236, 17076808, 17825791, 18574774, 19323757, 20072739, 20821722, 21570705, 22319688, 23068671, 23817654, 24566636, 25315619, 26064602, 26813585, 27562568, 28311551, 29060533, 29809516, 30558499, 31307482, 32056465, 32805448, 33554431}
-
-#define phaseModPWM_25 {0, 391991, 783982, 1175973, 1567964, 1959955, 2351946, 2743937, 3135928, 3527919, 3919910, 4311901, 4703892, 5095883, 5487874, 5879865, 6271856, 6663847, 7055838, 7447829, 7839820, 8231811, 8623802, 9015793, 9407784, 9799775, 10191766, 10583757, 10975748, 11367739, 11759730, 12151721, 12543712, 12935703, 13327694, 13719685, 14111676, 14503667, 14895658, 15287649, 15679640, 16071631, 16463622, 16935491, 17726869, 18518247, 19309625, 20101003, 20892381, 21683759, 22475137, 23266515, 24057893, 24849272, 25640650, 26432028, 27223406, 28014784, 28806162, 29597540, 30388918, 31180296, 31971674, 32763052, 33554431}
-
-#define phaseModPWM_26 {0, 381300, 762600, 1143901, 1525201, 1906501, 2287802, 2669102, 3050402, 3431703, 3813003, 4194303, 4575604, 4956904, 5338204, 5719505, 6100805, 6482105, 6863406, 7244706, 7626007, 8007307, 8388607, 8769908, 9151208, 9532508, 9913809, 10295109, 10676409, 11057710, 11439010, 11820310, 12201611, 12582911, 12964211, 13345512, 13726812, 14108113, 14489413, 14870713, 15252014, 15633314, 16014614, 16395915, 16777215, 17616076, 18454937, 19293797, 20132658, 20971519, 21810380, 22649240, 23488101, 24326962, 25165823, 26004684, 26843544, 27682405, 28521266, 29360127, 30198987, 31037848, 31876709, 32715570, 33554431}
-
-#define phaseModPWM_27 {0, 371177, 742354, 1113532, 1484709, 1855886, 2227064, 2598241, 2969418, 3340596, 3711773, 4082950, 4454128, 4825305, 5196482, 5567660, 5938837, 6310014, 6681192, 7052369, 7423546, 7794724, 8165901, 8537078, 8908256, 9279433, 9650610, 10021788, 10392965, 10764142, 11135320, 11506497, 11877674, 12248852, 12620029, 12991206, 13362384, 13733561, 14104738, 14475916, 14847093, 15218270, 15589448, 15960625, 16331802, 16702980, 17491139, 18383544, 19275949, 20168354, 21060759, 21953164, 22845570, 23737975, 24630380, 25522785, 26415190, 27307595, 28200000, 29092405, 29984810, 30877215, 31769620, 32662025, 33554431}
-
-#define phaseModPWM_28 {0, 361577, 723155, 1084733, 1446311, 1807889, 2169467, 2531045, 2892623, 3254201, 3615779, 3977357, 4338935, 4700512, 5062090, 5423668, 5785246, 6146824, 6508402, 6869980, 7231558, 7593136, 7954714, 8316292, 8677870, 9039448, 9401025, 9762603, 10124181, 10485759, 10847337, 11208915, 11570493, 11932071, 12293649, 12655227, 13016805, 13378383, 13739960, 14101538, 14463116, 14824694, 15186272, 15547850, 15909428, 16271006, 16632584, 17349166, 18302416, 19255667, 20208918, 21162169, 22115420, 23068671, 24021922, 24975173, 25928423, 26881674, 27834925, 28788176, 29741427, 30694678, 31647929, 32601180, 33554430}
-
-#define phaseModPWM_29 {0, 352462, 704925, 1057387, 1409850, 1762312, 2114775, 2467237, 2819700, 3172162, 3524625, 3877087, 4229550, 4582012, 4934475, 5286937, 5639400, 5991862, 6344325, 6696787, 7049250, 7401712, 7754175, 8106637, 8459100, 8811562, 9164025, 9516487, 9868950, 10221412, 10573875, 10926337, 11278800, 11631262, 11983725, 12336187, 12688650, 13041112, 13393575, 13746037, 14098500, 14450962, 14803425, 15155887, 15508350, 15860812, 16213275, 16565737, 17186415, 18209416, 19232417, 20255418, 21278419, 22301420, 23324421, 24347422, 25370423, 26393424, 27416425, 28439426, 29462427, 30485428, 31508429, 32531430, 33554431}
-
-#define phaseModPWM_30 {0, 343795, 687590, 1031386, 1375181, 1718976, 2062772, 2406567, 2750363, 3094158, 3437953, 3781749, 4125544, 4469340, 4813135, 5156930, 5500726, 5844521, 6188317, 6532112, 6875907, 7219703, 7563498, 7907294, 8251089, 8594884, 8938680, 9282475, 9626271, 9970066, 10313861, 10657657, 11001452, 11345248, 11689043, 12032838, 12376634, 12720429, 13064225, 13408020, 13751815, 14095611, 14439406, 14783202, 15126997, 15470792, 15814588, 16158383, 16502179, 16997968, 18101732, 19205496, 20309260, 21413025, 22516789, 23620553, 24724317, 25828081, 26931845, 28035610, 29139374, 30243138, 31346902, 32450666, 33554431}
-
-#define phaseModPWM_31 {0, 335544, 671088, 1006632, 1342177, 1677721, 2013265, 2348810, 2684354, 3019898, 3355443, 3690987, 4026531, 4362076, 4697620, 5033164, 5368708, 5704253, 6039797, 6375341, 6710886, 7046430, 7381974, 7717519, 8053063, 8388607, 8724152, 9059696, 9395240, 9730784, 10066329, 10401873, 10737417, 11072962, 11408506, 11744050, 12079595, 12415139, 12750683, 13086228, 13421772, 13757316, 14092861, 14428405, 14763949, 15099493, 15435038, 15770582, 16106126, 16441671, 16777215, 17975588, 19173960, 20372333, 21570705, 22769078, 23967450, 25165823, 26364195, 27562568, 28760940, 29959313, 31157685, 32356058, 33554431}
-
-#define phaseModPWM_32 {0, 327679, 655359, 983039, 1310719, 1638399, 1966079, 2293759, 2621439, 2949119, 3276799, 3604479, 3932159, 4259839, 4587519, 4915199, 5242879, 5570559, 5898239, 6225919, 6553599, 6881279, 7208959, 7536639, 7864319, 8191999, 8519679, 8847359, 9175039, 9502719, 9830399, 10158079, 10485759, 10813439, 11141119, 11468799, 11796479, 12124159, 12451839, 12779519, 13107199, 13434879, 13762559, 14090239, 14417919, 14745599, 15073279, 15400959, 15728639, 16056319, 16383999, 16711679, 17825791, 19136511, 20447231, 21757951, 23068671, 24379391, 25690111, 27000831, 28311551, 29622271, 30932991, 32243711, 33554431}
-
-
-
-/*
- *
- * Oscillators
- *
- */
-
-
-class Sine {
-
-
-public:
-
-	static constexpr int32_t big_sine[4097] = {16383, 16408, 16433, 16458, 16484, 16509, 16534, 16559, 16584, 16609, 16634, 16659, 16685, 16710, 16735, 16760, 16785, 16810, 16835, 16860, 16886, 16911, 16936, 16961, 16986, 17011, 17036, 17061, 17086, 17112, 17137, 17162, 17187, 17212, 17237, 17262, 17287, 17312, 17337, 17363, 17388, 17413, 17438, 17463, 17488, 17513, 17538, 17563, 17588, 17613, 17638, 17663, 17688, 17714, 17739, 17764, 17789, 17814, 17839, 17864, 17889, 17914, 17939, 17964, 17989, 18014, 18039, 18064, 18089, 18114, 18139, 18164, 18189, 18214, 18239, 18264, 18289, 18314, 18339, 18364, 18389, 18413, 18438, 18463, 18488, 18513, 18538, 18563, 18588, 18613, 18638, 18663, 18687, 18712, 18737, 18762, 18787, 18812, 18837, 18862, 18886, 18911, 18936, 18961, 18986, 19010, 19035, 19060, 19085, 19110, 19134, 19159, 19184, 19209, 19233, 19258, 19283, 19308, 19332, 19357, 19382, 19407, 19431, 19456, 19481, 19505, 19530, 19555, 19579, 19604, 19629, 19653, 19678, 19702, 19727, 19752, 19776, 19801, 19825, 19850, 19874, 19899, 19924, 19948, 19973, 19997, 20022, 20046, 20071, 20095, 20120, 20144, 20169, 20193, 20217, 20242, 20266, 20291, 20315, 20339, 20364, 20388, 20413, 20437, 20461, 20486, 20510, 20534, 20559, 20583, 20607, 20631, 20656, 20680, 20704, 20728, 20753, 20777, 20801, 20825, 20849, 20874, 20898, 20922, 20946, 20970, 20994, 21018, 21043, 21067, 21091, 21115, 21139, 21163, 21187, 21211, 21235, 21259, 21283, 21307, 21331, 21355, 21379, 21403, 21427, 21451, 21474, 21498, 21522, 21546, 21570, 21594, 21618, 21641, 21665, 21689, 21713, 21736, 21760, 21784, 21808, 21831, 21855, 21879, 21902, 21926, 21950, 21973, 21997, 22021, 22044, 22068, 22091, 22115, 22138, 22162, 22185, 22209, 22232, 22256, 22279, 22303, 22326, 22350, 22373, 22396, 22420, 22443, 22466, 22490, 22513, 22536, 22560, 22583, 22606, 22629, 22653, 22676, 22699, 22722, 22745, 22769, 22792, 22815, 22838, 22861, 22884, 22907, 22930, 22953, 22976, 22999, 23022, 23045, 23068, 23091, 23114, 23137, 23160, 23183, 23206, 23228, 23251, 23274, 23297, 23320, 23342, 23365, 23388, 23411, 23433, 23456, 23479, 23501, 23524, 23546, 23569, 23592, 23614, 23637, 23659, 23682, 23704, 23727, 23749, 23772, 23794, 23816, 23839, 23861, 23884, 23906, 23928, 23951, 23973, 23995, 24017, 24040, 24062, 24084, 24106, 24128, 24150, 24173, 24195, 24217, 24239, 24261, 24283, 24305, 24327, 24349, 24371, 24393, 24415, 24437, 24458, 24480, 24502, 24524, 24546, 24567, 24589, 24611, 24633, 24654, 24676, 24698, 24719, 24741, 24763, 24784, 24806, 24827, 24849, 24870, 24892, 24913, 24935, 24956, 24978, 24999, 25020, 25042, 25063, 25084, 25106, 25127, 25148, 25169, 25191, 25212, 25233, 25254, 25275, 25296, 25317, 25338, 25359, 25380, 25401, 25422, 25443, 25464, 25485, 25506, 25527, 25548, 25569, 25589, 25610, 25631, 25652, 25672, 25693, 25714, 25734, 25755, 25776, 25796, 25817, 25837, 25858, 25878, 25899, 25919, 25940, 25960, 25980, 26001, 26021, 26041, 26062, 26082, 26102, 26122, 26143, 26163, 26183, 26203, 26223, 26243, 26263, 26283, 26303, 26323, 26343, 26363, 26383, 26403, 26423, 26443, 26463, 26482, 26502, 26522, 26542, 26561, 26581, 26601, 26620, 26640, 26660, 26679, 26699, 26718, 26738, 26757, 26777, 26796, 26815, 26835, 26854, 26873, 26893, 26912, 26931, 26950, 26970, 26989, 27008, 27027, 27046, 27065, 27084, 27103, 27122, 27141, 27160, 27179, 27198, 27217, 27236, 27255, 27273, 27292, 27311, 27330, 27348, 27367, 27385, 27404, 27423, 27441, 27460, 27478, 27497, 27515, 27534, 27552, 27570, 27589, 27607, 27625, 27644, 27662, 27680, 27698, 27716, 27735, 27753, 27771, 27789, 27807, 27825, 27843, 27861, 27879, 27897, 27914, 27932, 27950, 27968, 27986, 28003, 28021, 28039, 28056, 28074, 28092, 28109, 28127, 28144, 28162, 28179, 28197, 28214, 28231, 28249, 28266, 28283, 28301, 28318, 28335, 28352, 28369, 28386, 28404, 28421, 28438, 28455, 28472, 28489, 28505, 28522, 28539, 28556, 28573, 28590, 28606, 28623, 28640, 28656, 28673, 28690, 28706, 28723, 28739, 28756, 28772, 28789, 28805, 28822, 28838, 28854, 28870, 28887, 28903, 28919, 28935, 28951, 28968, 28984, 29000, 29016, 29032, 29048, 29064, 29079, 29095, 29111, 29127, 29143, 29158, 29174, 29190, 29206, 29221, 29237, 29252, 29268, 29283, 29299, 29314, 29330, 29345, 29360, 29376, 29391, 29406, 29422, 29437, 29452, 29467, 29482, 29497, 29512, 29527, 29542, 29557, 29572, 29587, 29602, 29617, 29632, 29646, 29661, 29676, 29691, 29705, 29720, 29734, 29749, 29763, 29778, 29792, 29807, 29821, 29836, 29850, 29864, 29878, 29893, 29907, 29921, 29935, 29949, 29963, 29977, 29991, 30005, 30019, 30033, 30047, 30061, 30075, 30089, 30102, 30116, 30130, 30143, 30157, 30171, 30184, 30198, 30211, 30225, 30238, 30251, 30265, 30278, 30291, 30305, 30318, 30331, 30344, 30357, 30371, 30384, 30397, 30410, 30423, 30436, 30449, 30461, 30474, 30487, 30500, 30513, 30525, 30538, 30551, 30563, 30576, 30588, 30601, 30613, 30626, 30638, 30650, 30663, 30675, 30687, 30700, 30712, 30724, 30736, 30748, 30760, 30772, 30784, 30796, 30808, 30820, 30832, 30844, 30856, 30867, 30879, 30891, 30902, 30914, 30926, 30937, 30949, 30960, 30972, 30983, 30994, 31006, 31017, 31028, 31040, 31051, 31062, 31073, 31084, 31095, 31106, 31117, 31128, 31139, 31150, 31161, 31172, 31183, 31194, 31204, 31215, 31226, 31236, 31247, 31257, 31268, 31278, 31289, 31299, 31310, 31320, 31330, 31341, 31351, 31361, 31371, 31381, 31391, 31401, 31411, 31421, 31431, 31441, 31451, 31461, 31471, 31481, 31490, 31500, 31510, 31519, 31529, 31539, 31548, 31558, 31567, 31576, 31586, 31595, 31604, 31614, 31623, 31632, 31641, 31651, 31660, 31669, 31678, 31687, 31696, 31705, 31713, 31722, 31731, 31740, 31749, 31757, 31766, 31775, 31783, 31792, 31800, 31809, 31817, 31826, 31834, 31842, 31851, 31859, 31867, 31875, 31884, 31892, 31900, 31908, 31916, 31924, 31932, 31940, 31947, 31955, 31963, 31971, 31979, 31986, 31994, 32001, 32009, 32017, 32024, 32032, 32039, 32046, 32054, 32061, 32068, 32076, 32083, 32090, 32097, 32104, 32111, 32118, 32125, 32132, 32139, 32146, 32153, 32160, 32166, 32173, 32180, 32186, 32193, 32200, 32206, 32213, 32219, 32225, 32232, 32238, 32245, 32251, 32257, 32263, 32269, 32276, 32282, 32288, 32294, 32300, 32306, 32311, 32317, 32323, 32329, 32335, 32340, 32346, 32352, 32357, 32363, 32368, 32374, 32379, 32385, 32390, 32395, 32401, 32406, 32411, 32416, 32422, 32427, 32432, 32437, 32442, 32447, 32452, 32457, 32461, 32466, 32471, 32476, 32480, 32485, 32490, 32494, 32499, 32503, 32508, 32512, 32517, 32521, 32525, 32530, 32534, 32538, 32542, 32546, 32550, 32554, 32558, 32562, 32566, 32570, 32574, 32578, 32582, 32585, 32589, 32593, 32596, 32600, 32604, 32607, 32611, 32614, 32617, 32621, 32624, 32627, 32631, 32634, 32637, 32640, 32643, 32646, 32649, 32652, 32655, 32658, 32661, 32664, 32667, 32669, 32672, 32675, 32677, 32680, 32683, 32685, 32688, 32690, 32692, 32695, 32697, 32699, 32702, 32704, 32706, 32708, 32710, 32712, 32714, 32716, 32718, 32720, 32722, 32724, 32726, 32727, 32729, 32731, 32733, 32734, 32736, 32737, 32739, 32740, 32742, 32743, 32744, 32746, 32747, 32748, 32749, 32750, 32751, 32752, 32753, 32754, 32755, 32756, 32757, 32758, 32759, 32760, 32760, 32761, 32762, 32762, 32763, 32763, 32764, 32764, 32765, 32765, 32765, 32766, 32766, 32766, 32766, 32766, 32766, 32766, 32767, 32766, 32766, 32766, 32766, 32766, 32766, 32766, 32765, 32765, 32765, 32764, 32764, 32763, 32763, 32762, 32762, 32761, 32760, 32760, 32759, 32758, 32757, 32756, 32755, 32754, 32753, 32752, 32751, 32750, 32749, 32748, 32747, 32746, 32744, 32743, 32742, 32740, 32739, 32737, 32736, 32734, 32733, 32731, 32729, 32727, 32726, 32724, 32722, 32720, 32718, 32716, 32714, 32712, 32710, 32708, 32706, 32704, 32702, 32699, 32697, 32695, 32692, 32690, 32688, 32685, 32683, 32680, 32677, 32675, 32672, 32669, 32667, 32664, 32661, 32658, 32655, 32652, 32649, 32646, 32643, 32640, 32637, 32634, 32631, 32627, 32624, 32621, 32617, 32614, 32611, 32607, 32604, 32600, 32596, 32593, 32589, 32585, 32582, 32578, 32574, 32570, 32566, 32562, 32558, 32554, 32550, 32546, 32542, 32538, 32534, 32530, 32525, 32521, 32517, 32512, 32508, 32503, 32499, 32494, 32490, 32485, 32480, 32476, 32471, 32466, 32461, 32457, 32452, 32447, 32442, 32437, 32432, 32427, 32422, 32416, 32411, 32406, 32401, 32395, 32390, 32385, 32379, 32374, 32368, 32363, 32357, 32352, 32346, 32340, 32335, 32329, 32323, 32317, 32311, 32306, 32300, 32294, 32288, 32282, 32276, 32269, 32263, 32257, 32251, 32245, 32238, 32232, 32225, 32219, 32213, 32206, 32200, 32193, 32186, 32180, 32173, 32166, 32160, 32153, 32146, 32139, 32132, 32125, 32118, 32111, 32104, 32097, 32090, 32083, 32076, 32068, 32061, 32054, 32046, 32039, 32032, 32024, 32017, 32009, 32001, 31994, 31986, 31979, 31971, 31963, 31955, 31947, 31940, 31932, 31924, 31916, 31908, 31900, 31892, 31884, 31875, 31867, 31859, 31851, 31842, 31834, 31826, 31817, 31809, 31800, 31792, 31783, 31775, 31766, 31757, 31749, 31740, 31731, 31722, 31713, 31705, 31696, 31687, 31678, 31669, 31660, 31651, 31641, 31632, 31623, 31614, 31604, 31595, 31586, 31576, 31567, 31558, 31548, 31539, 31529, 31519, 31510, 31500, 31490, 31481, 31471, 31461, 31451, 31441, 31431, 31421, 31411, 31401, 31391, 31381, 31371, 31361, 31351, 31341, 31330, 31320, 31310, 31299, 31289, 31278, 31268, 31257, 31247, 31236, 31226, 31215, 31204, 31194, 31183, 31172, 31161, 31150, 31139, 31128, 31117, 31106, 31095, 31084, 31073, 31062, 31051, 31040, 31028, 31017, 31006, 30994, 30983, 30972, 30960, 30949, 30937, 30926, 30914, 30902, 30891, 30879, 30867, 30856, 30844, 30832, 30820, 30808, 30796, 30784, 30772, 30760, 30748, 30736, 30724, 30712, 30700, 30687, 30675, 30663, 30650, 30638, 30626, 30613, 30601, 30588, 30576, 30563, 30551, 30538, 30525, 30513, 30500, 30487, 30474, 30461, 30449, 30436, 30423, 30410, 30397, 30384, 30371, 30357, 30344, 30331, 30318, 30305, 30291, 30278, 30265, 30251, 30238, 30225, 30211, 30198, 30184, 30171, 30157, 30143, 30130, 30116, 30102, 30089, 30075, 30061, 30047, 30033, 30019, 30005, 29991, 29977, 29963, 29949, 29935, 29921, 29907, 29893, 29878, 29864, 29850, 29836, 29821, 29807, 29792, 29778, 29763, 29749, 29734, 29720, 29705, 29691, 29676, 29661, 29646, 29632, 29617, 29602, 29587, 29572, 29557, 29542, 29527, 29512, 29497, 29482, 29467, 29452, 29437, 29422, 29406, 29391, 29376, 29360, 29345, 29330, 29314, 29299, 29283, 29268, 29252, 29237, 29221, 29206, 29190, 29174, 29158, 29143, 29127, 29111, 29095, 29079, 29064, 29048, 29032, 29016, 29000, 28984, 28968, 28951, 28935, 28919, 28903, 28887, 28870, 28854, 28838, 28822, 28805, 28789, 28772, 28756, 28739, 28723, 28706, 28690, 28673, 28656, 28640, 28623, 28606, 28590, 28573, 28556, 28539, 28522, 28505, 28489, 28472, 28455, 28438, 28421, 28404, 28386, 28369, 28352, 28335, 28318, 28301, 28283, 28266, 28249, 28231, 28214, 28197, 28179, 28162, 28144, 28127, 28109, 28092, 28074, 28056, 28039, 28021, 28003, 27986, 27968, 27950, 27932, 27914, 27897, 27879, 27861, 27843, 27825, 27807, 27789, 27771, 27753, 27735, 27716, 27698, 27680, 27662, 27644, 27625, 27607, 27589, 27570, 27552, 27534, 27515, 27497, 27478, 27460, 27441, 27423, 27404, 27385, 27367, 27348, 27330, 27311, 27292, 27273, 27255, 27236, 27217, 27198, 27179, 27160, 27141, 27122, 27103, 27084, 27065, 27046, 27027, 27008, 26989, 26970, 26950, 26931, 26912, 26893, 26873, 26854, 26835, 26815, 26796, 26777, 26757, 26738, 26718, 26699, 26679, 26660, 26640, 26620, 26601, 26581, 26561, 26542, 26522, 26502, 26482, 26463, 26443, 26423, 26403, 26383, 26363, 26343, 26323, 26303, 26283, 26263, 26243, 26223, 26203, 26183, 26163, 26143, 26122, 26102, 26082, 26062, 26041, 26021, 26001, 25980, 25960, 25940, 25919, 25899, 25878, 25858, 25837, 25817, 25796, 25776, 25755, 25734, 25714, 25693, 25672, 25652, 25631, 25610, 25589, 25569, 25548, 25527, 25506, 25485, 25464, 25443, 25422, 25401, 25380, 25359, 25338, 25317, 25296, 25275, 25254, 25233, 25212, 25191, 25169, 25148, 25127, 25106, 25084, 25063, 25042, 25020, 24999, 24978, 24956, 24935, 24913, 24892, 24870, 24849, 24827, 24806, 24784, 24763, 24741, 24719, 24698, 24676, 24654, 24633, 24611, 24589, 24567, 24546, 24524, 24502, 24480, 24458, 24437, 24415, 24393, 24371, 24349, 24327, 24305, 24283, 24261, 24239, 24217, 24195, 24173, 24150, 24128, 24106, 24084, 24062, 24040, 24017, 23995, 23973, 23951, 23928, 23906, 23884, 23861, 23839, 23816, 23794, 23772, 23749, 23727, 23704, 23682, 23659, 23637, 23614, 23592, 23569, 23546, 23524, 23501, 23479, 23456, 23433, 23411, 23388, 23365, 23342, 23320, 23297, 23274, 23251, 23228, 23206, 23183, 23160, 23137, 23114, 23091, 23068, 23045, 23022, 22999, 22976, 22953, 22930, 22907, 22884, 22861, 22838, 22815, 22792, 22769, 22745, 22722, 22699, 22676, 22653, 22629, 22606, 22583, 22560, 22536, 22513, 22490, 22466, 22443, 22420, 22396, 22373, 22350, 22326, 22303, 22279, 22256, 22232, 22209, 22185, 22162, 22138, 22115, 22091, 22068, 22044, 22021, 21997, 21973, 21950, 21926, 21902, 21879, 21855, 21831, 21808, 21784, 21760, 21736, 21713, 21689, 21665, 21641, 21618, 21594, 21570, 21546, 21522, 21498, 21474, 21451, 21427, 21403, 21379, 21355, 21331, 21307, 21283, 21259, 21235, 21211, 21187, 21163, 21139, 21115, 21091, 21067, 21043, 21018, 20994, 20970, 20946, 20922, 20898, 20874, 20849, 20825, 20801, 20777, 20753, 20728, 20704, 20680, 20656, 20631, 20607, 20583, 20559, 20534, 20510, 20486, 20461, 20437, 20413, 20388, 20364, 20339, 20315, 20291, 20266, 20242, 20217, 20193, 20169, 20144, 20120, 20095, 20071, 20046, 20022, 19997, 19973, 19948, 19924, 19899, 19874, 19850, 19825, 19801, 19776, 19752, 19727, 19702, 19678, 19653, 19629, 19604, 19579, 19555, 19530, 19505, 19481, 19456, 19431, 19407, 19382, 19357, 19332, 19308, 19283, 19258, 19233, 19209, 19184, 19159, 19134, 19110, 19085, 19060, 19035, 19010, 18986, 18961, 18936, 18911, 18886, 18862, 18837, 18812, 18787, 18762, 18737, 18712, 18687, 18663, 18638, 18613, 18588, 18563, 18538, 18513, 18488, 18463, 18438, 18413, 18389, 18364, 18339, 18314, 18289, 18264, 18239, 18214, 18189, 18164, 18139, 18114, 18089, 18064, 18039, 18014, 17989, 17964, 17939, 17914, 17889, 17864, 17839, 17814, 17789, 17764, 17739, 17714, 17688, 17663, 17638, 17613, 17588, 17563, 17538, 17513, 17488, 17463, 17438, 17413, 17388, 17363, 17337, 17312, 17287, 17262, 17237, 17212, 17187, 17162, 17137, 17112, 17086, 17061, 17036, 17011, 16986, 16961, 16936, 16911, 16886, 16860, 16835, 16810, 16785, 16760, 16735, 16710, 16685, 16659, 16634, 16609, 16584, 16559, 16534, 16509, 16484, 16458, 16433, 16408, 16383, 16358, 16333, 16308, 16282, 16257, 16232, 16207, 16182, 16157, 16132, 16107, 16081, 16056, 16031, 16006, 15981, 15956, 15931, 15906, 15880, 15855, 15830, 15805, 15780, 15755, 15730, 15705, 15680, 15654, 15629, 15604, 15579, 15554, 15529, 15504, 15479, 15454, 15429, 15403, 15378, 15353, 15328, 15303, 15278, 15253, 15228, 15203, 15178, 15153, 15128, 15103, 15078, 15052, 15027, 15002, 14977, 14952, 14927, 14902, 14877, 14852, 14827, 14802, 14777, 14752, 14727, 14702, 14677, 14652, 14627, 14602, 14577, 14552, 14527, 14502, 14477, 14452, 14427, 14402, 14377, 14353, 14328, 14303, 14278, 14253, 14228, 14203, 14178, 14153, 14128, 14103, 14079, 14054, 14029, 14004, 13979, 13954, 13929, 13904, 13880, 13855, 13830, 13805, 13780, 13756, 13731, 13706, 13681, 13656, 13632, 13607, 13582, 13557, 13533, 13508, 13483, 13458, 13434, 13409, 13384, 13359, 13335, 13310, 13285, 13261, 13236, 13211, 13187, 13162, 13137, 13113, 13088, 13064, 13039, 13014, 12990, 12965, 12941, 12916, 12892, 12867, 12842, 12818, 12793, 12769, 12744, 12720, 12695, 12671, 12646, 12622, 12597, 12573, 12549, 12524, 12500, 12475, 12451, 12427, 12402, 12378, 12353, 12329, 12305, 12280, 12256, 12232, 12207, 12183, 12159, 12135, 12110, 12086, 12062, 12038, 12013, 11989, 11965, 11941, 11917, 11892, 11868, 11844, 11820, 11796, 11772, 11748, 11723, 11699, 11675, 11651, 11627, 11603, 11579, 11555, 11531, 11507, 11483, 11459, 11435, 11411, 11387, 11363, 11339, 11315, 11292, 11268, 11244, 11220, 11196, 11172, 11148, 11125, 11101, 11077, 11053, 11030, 11006, 10982, 10958, 10935, 10911, 10887, 10864, 10840, 10816, 10793, 10769, 10745, 10722, 10698, 10675, 10651, 10628, 10604, 10581, 10557, 10534, 10510, 10487, 10463, 10440, 10416, 10393, 10370, 10346, 10323, 10300, 10276, 10253, 10230, 10206, 10183, 10160, 10137, 10113, 10090, 10067, 10044, 10021, 9997, 9974, 9951, 9928, 9905, 9882, 9859, 9836, 9813, 9790, 9767, 9744, 9721, 9698, 9675, 9652, 9629, 9606, 9583, 9560, 9538, 9515, 9492, 9469, 9446, 9424, 9401, 9378, 9355, 9333, 9310, 9287, 9265, 9242, 9220, 9197, 9174, 9152, 9129, 9107, 9084, 9062, 9039, 9017, 8994, 8972, 8950, 8927, 8905, 8882, 8860, 8838, 8815, 8793, 8771, 8749, 8726, 8704, 8682, 8660, 8638, 8616, 8593, 8571, 8549, 8527, 8505, 8483, 8461, 8439, 8417, 8395, 8373, 8351, 8329, 8308, 8286, 8264, 8242, 8220, 8199, 8177, 8155, 8133, 8112, 8090, 8068, 8047, 8025, 8003, 7982, 7960, 7939, 7917, 7896, 7874, 7853, 7831, 7810, 7788, 7767, 7746, 7724, 7703, 7682, 7660, 7639, 7618, 7597, 7575, 7554, 7533, 7512, 7491, 7470, 7449, 7428, 7407, 7386, 7365, 7344, 7323, 7302, 7281, 7260, 7239, 7218, 7197, 7177, 7156, 7135, 7114, 7094, 7073, 7052, 7032, 7011, 6990, 6970, 6949, 6929, 6908, 6888, 6867, 6847, 6826, 6806, 6786, 6765, 6745, 6725, 6704, 6684, 6664, 6644, 6623, 6603, 6583, 6563, 6543, 6523, 6503, 6483, 6463, 6443, 6423, 6403, 6383, 6363, 6343, 6323, 6303, 6284, 6264, 6244, 6224, 6205, 6185, 6165, 6146, 6126, 6106, 6087, 6067, 6048, 6028, 6009, 5989, 5970, 5951, 5931, 5912, 5893, 5873, 5854, 5835, 5816, 5796, 5777, 5758, 5739, 5720, 5701, 5682, 5663, 5644, 5625, 5606, 5587, 5568, 5549, 5530, 5511, 5493, 5474, 5455, 5436, 5418, 5399, 5381, 5362, 5343, 5325, 5306, 5288, 5269, 5251, 5232, 5214, 5196, 5177, 5159, 5141, 5122, 5104, 5086, 5068, 5050, 5031, 5013, 4995, 4977, 4959, 4941, 4923, 4905, 4887, 4869, 4852, 4834, 4816, 4798, 4780, 4763, 4745, 4727, 4710, 4692, 4674, 4657, 4639, 4622, 4604, 4587, 4569, 4552, 4535, 4517, 4500, 4483, 4465, 4448, 4431, 4414, 4397, 4380, 4362, 4345, 4328, 4311, 4294, 4277, 4261, 4244, 4227, 4210, 4193, 4176, 4160, 4143, 4126, 4110, 4093, 4076, 4060, 4043, 4027, 4010, 3994, 3977, 3961, 3944, 3928, 3912, 3896, 3879, 3863, 3847, 3831, 3815, 3798, 3782, 3766, 3750, 3734, 3718, 3702, 3687, 3671, 3655, 3639, 3623, 3608, 3592, 3576, 3560, 3545, 3529, 3514, 3498, 3483, 3467, 3452, 3436, 3421, 3406, 3390, 3375, 3360, 3344, 3329, 3314, 3299, 3284, 3269, 3254, 3239, 3224, 3209, 3194, 3179, 3164, 3149, 3134, 3120, 3105, 3090, 3075, 3061, 3046, 3032, 3017, 3003, 2988, 2974, 2959, 2945, 2930, 2916, 2902, 2888, 2873, 2859, 2845, 2831, 2817, 2803, 2789, 2775, 2761, 2747, 2733, 2719, 2705, 2691, 2677, 2664, 2650, 2636, 2623, 2609, 2595, 2582, 2568, 2555, 2541, 2528, 2515, 2501, 2488, 2475, 2461, 2448, 2435, 2422, 2409, 2395, 2382, 2369, 2356, 2343, 2330, 2317, 2305, 2292, 2279, 2266, 2253, 2241, 2228, 2215, 2203, 2190, 2178, 2165, 2153, 2140, 2128, 2116, 2103, 2091, 2079, 2066, 2054, 2042, 2030, 2018, 2006, 1994, 1982, 1970, 1958, 1946, 1934, 1922, 1910, 1899, 1887, 1875, 1864, 1852, 1840, 1829, 1817, 1806, 1794, 1783, 1772, 1760, 1749, 1738, 1726, 1715, 1704, 1693, 1682, 1671, 1660, 1649, 1638, 1627, 1616, 1605, 1594, 1583, 1572, 1562, 1551, 1540, 1530, 1519, 1509, 1498, 1488, 1477, 1467, 1456, 1446, 1436, 1425, 1415, 1405, 1395, 1385, 1375, 1365, 1355, 1345, 1335, 1325, 1315, 1305, 1295, 1285, 1276, 1266, 1256, 1247, 1237, 1227, 1218, 1208, 1199, 1190, 1180, 1171, 1162, 1152, 1143, 1134, 1125, 1115, 1106, 1097, 1088, 1079, 1070, 1061, 1053, 1044, 1035, 1026, 1017, 1009, 1000, 991, 983, 974, 966, 957, 949, 940, 932, 924, 915, 907, 899, 891, 882, 874, 866, 858, 850, 842, 834, 826, 819, 811, 803, 795, 787, 780, 772, 765, 757, 749, 742, 734, 727, 720, 712, 705, 698, 690, 683, 676, 669, 662, 655, 648, 641, 634, 627, 620, 613, 606, 600, 593, 586, 580, 573, 566, 560, 553, 547, 541, 534, 528, 521, 515, 509, 503, 497, 490, 484, 478, 472, 466, 460, 455, 449, 443, 437, 431, 426, 420, 414, 409, 403, 398, 392, 387, 381, 376, 371, 365, 360, 355, 350, 344, 339, 334, 329, 324, 319, 314, 309, 305, 300, 295, 290, 286, 281, 276, 272, 267, 263, 258, 254, 249, 245, 241, 236, 232, 228, 224, 220, 216, 212, 208, 204, 200, 196, 192, 188, 184, 181, 177, 173, 170, 166, 162, 159, 155, 152, 149, 145, 142, 139, 135, 132, 129, 126, 123, 120, 117, 114, 111, 108, 105, 102, 99, 97, 94, 91, 89, 86, 83, 81, 78, 76, 74, 71, 69, 67, 64, 62, 60, 58, 56, 54, 52, 50, 48, 46, 44, 42, 40, 39, 37, 35, 33, 32, 30, 29, 27, 26, 24, 23, 22, 20, 19, 18, 17, 16, 15, 14, 13, 12, 11, 10, 9, 8, 7, 6, 6, 5, 4, 4, 3, 3, 2, 2, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 2, 2, 3, 3, 4, 4, 5, 6, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20, 22, 23, 24, 26, 27, 29, 30, 32, 33, 35, 37, 39, 40, 42, 44, 46, 48, 50, 52, 54, 56, 58, 60, 62, 64, 67, 69, 71, 74, 76, 78, 81, 83, 86, 89, 91, 94, 97, 99, 102, 105, 108, 111, 114, 117, 120, 123, 126, 129, 132, 135, 139, 142, 145, 149, 152, 155, 159, 162, 166, 170, 173, 177, 181, 184, 188, 192, 196, 200, 204, 208, 212, 216, 220, 224, 228, 232, 236, 241, 245, 249, 254, 258, 263, 267, 272, 276, 281, 286, 290, 295, 300, 305, 309, 314, 319, 324, 329, 334, 339, 344, 350, 355, 360, 365, 371, 376, 381, 387, 392, 398, 403, 409, 414, 420, 426, 431, 437, 443, 449, 455, 460, 466, 472, 478, 484, 490, 497, 503, 509, 515, 521, 528, 534, 541, 547, 553, 560, 566, 573, 580, 586, 593, 600, 606, 613, 620, 627, 634, 641, 648, 655, 662, 669, 676, 683, 690, 698, 705, 712, 720, 727, 734, 742, 749, 757, 765, 772, 780, 787, 795, 803, 811, 819, 826, 834, 842, 850, 858, 866, 874, 882, 891, 899, 907, 915, 924, 932, 940, 949, 957, 966, 974, 983, 991, 1000, 1009, 1017, 1026, 1035, 1044, 1053, 1061, 1070, 1079, 1088, 1097, 1106, 1115, 1125, 1134, 1143, 1152, 1162, 1171, 1180, 1190, 1199, 1208, 1218, 1227, 1237, 1247, 1256, 1266, 1276, 1285, 1295, 1305, 1315, 1325, 1335, 1345, 1355, 1365, 1375, 1385, 1395, 1405, 1415, 1425, 1436, 1446, 1456, 1467, 1477, 1488, 1498, 1509, 1519, 1530, 1540, 1551, 1562, 1572, 1583, 1594, 1605, 1616, 1627, 1638, 1649, 1660, 1671, 1682, 1693, 1704, 1715, 1726, 1738, 1749, 1760, 1772, 1783, 1794, 1806, 1817, 1829, 1840, 1852, 1864, 1875, 1887, 1899, 1910, 1922, 1934, 1946, 1958, 1970, 1982, 1994, 2006, 2018, 2030, 2042, 2054, 2066, 2079, 2091, 2103, 2116, 2128, 2140, 2153, 2165, 2178, 2190, 2203, 2215, 2228, 2241, 2253, 2266, 2279, 2292, 2305, 2317, 2330, 2343, 2356, 2369, 2382, 2395, 2409, 2422, 2435, 2448, 2461, 2475, 2488, 2501, 2515, 2528, 2541, 2555, 2568, 2582, 2595, 2609, 2623, 2636, 2650, 2664, 2677, 2691, 2705, 2719, 2733, 2747, 2761, 2775, 2789, 2803, 2817, 2831, 2845, 2859, 2873, 2888, 2902, 2916, 2930, 2945, 2959, 2974, 2988, 3003, 3017, 3032, 3046, 3061, 3075, 3090, 3105, 3120, 3134, 3149, 3164, 3179, 3194, 3209, 3224, 3239, 3254, 3269, 3284, 3299, 3314, 3329, 3344, 3360, 3375, 3390, 3406, 3421, 3436, 3452, 3467, 3483, 3498, 3514, 3529, 3545, 3560, 3576, 3592, 3608, 3623, 3639, 3655, 3671, 3687, 3702, 3718, 3734, 3750, 3766, 3782, 3798, 3815, 3831, 3847, 3863, 3879, 3896, 3912, 3928, 3944, 3961, 3977, 3994, 4010, 4027, 4043, 4060, 4076, 4093, 4110, 4126, 4143, 4160, 4176, 4193, 4210, 4227, 4244, 4261, 4277, 4294, 4311, 4328, 4345, 4362, 4380, 4397, 4414, 4431, 4448, 4465, 4483, 4500, 4517, 4535, 4552, 4569, 4587, 4604, 4622, 4639, 4657, 4674, 4692, 4710, 4727, 4745, 4763, 4780, 4798, 4816, 4834, 4852, 4869, 4887, 4905, 4923, 4941, 4959, 4977, 4995, 5013, 5031, 5050, 5068, 5086, 5104, 5122, 5141, 5159, 5177, 5196, 5214, 5232, 5251, 5269, 5288, 5306, 5325, 5343, 5362, 5381, 5399, 5418, 5436, 5455, 5474, 5493, 5511, 5530, 5549, 5568, 5587, 5606, 5625, 5644, 5663, 5682, 5701, 5720, 5739, 5758, 5777, 5796, 5816, 5835, 5854, 5873, 5893, 5912, 5931, 5951, 5970, 5989, 6009, 6028, 6048, 6067, 6087, 6106, 6126, 6146, 6165, 6185, 6205, 6224, 6244, 6264, 6284, 6303, 6323, 6343, 6363, 6383, 6403, 6423, 6443, 6463, 6483, 6503, 6523, 6543, 6563, 6583, 6603, 6623, 6644, 6664, 6684, 6704, 6725, 6745, 6765, 6786, 6806, 6826, 6847, 6867, 6888, 6908, 6929, 6949, 6970, 6990, 7011, 7032, 7052, 7073, 7094, 7114, 7135, 7156, 7177, 7197, 7218, 7239, 7260, 7281, 7302, 7323, 7344, 7365, 7386, 7407, 7428, 7449, 7470, 7491, 7512, 7533, 7554, 7575, 7597, 7618, 7639, 7660, 7682, 7703, 7724, 7746, 7767, 7788, 7810, 7831, 7853, 7874, 7896, 7917, 7939, 7960, 7982, 8003, 8025, 8047, 8068, 8090, 8112, 8133, 8155, 8177, 8199, 8220, 8242, 8264, 8286, 8308, 8329, 8351, 8373, 8395, 8417, 8439, 8461, 8483, 8505, 8527, 8549, 8571, 8593, 8616, 8638, 8660, 8682, 8704, 8726, 8749, 8771, 8793, 8815, 8838, 8860, 8882, 8905, 8927, 8950, 8972, 8994, 9017, 9039, 9062, 9084, 9107, 9129, 9152, 9174, 9197, 9220, 9242, 9265, 9287, 9310, 9333, 9355, 9378, 9401, 9424, 9446, 9469, 9492, 9515, 9538, 9560, 9583, 9606, 9629, 9652, 9675, 9698, 9721, 9744, 9767, 9790, 9813, 9836, 9859, 9882, 9905, 9928, 9951, 9974, 9997, 10021, 10044, 10067, 10090, 10113, 10137, 10160, 10183, 10206, 10230, 10253, 10276, 10300, 10323, 10346, 10370, 10393, 10416, 10440, 10463, 10487, 10510, 10534, 10557, 10581, 10604, 10628, 10651, 10675, 10698, 10722, 10745, 10769, 10793, 10816, 10840, 10864, 10887, 10911, 10935, 10958, 10982, 11006, 11030, 11053, 11077, 11101, 11125, 11148, 11172, 11196, 11220, 11244, 11268, 11292, 11315, 11339, 11363, 11387, 11411, 11435, 11459, 11483, 11507, 11531, 11555, 11579, 11603, 11627, 11651, 11675, 11699, 11723, 11748, 11772, 11796, 11820, 11844, 11868, 11892, 11917, 11941, 11965, 11989, 12013, 12038, 12062, 12086, 12110, 12135, 12159, 12183, 12207, 12232, 12256, 12280, 12305, 12329, 12353, 12378, 12402, 12427, 12451, 12475, 12500, 12524, 12549, 12573, 12597, 12622, 12646, 12671, 12695, 12720, 12744, 12769, 12793, 12818, 12842, 12867, 12892, 12916, 12941, 12965, 12990, 13014, 13039, 13064, 13088, 13113, 13137, 13162, 13187, 13211, 13236, 13261, 13285, 13310, 13335, 13359, 13384, 13409, 13434, 13458, 13483, 13508, 13533, 13557, 13582, 13607, 13632, 13656, 13681, 13706, 13731, 13756, 13780, 13805, 13830, 13855, 13880, 13904, 13929, 13954, 13979, 14004, 14029, 14054, 14079, 14103, 14128, 14153, 14178, 14203, 14228, 14253, 14278, 14303, 14328, 14353, 14377, 14402, 14427, 14452, 14477, 14502, 14527, 14552, 14577, 14602, 14627, 14652, 14677, 14702, 14727, 14752, 14777, 14802, 14827, 14852, 14877, 14902, 14927, 14952, 14977, 15002, 15027, 15052, 15078, 15103, 15128, 15153, 15178, 15203, 15228, 15253, 15278, 15303, 15328, 15353, 15378, 15403, 15429, 15454, 15479, 15504, 15529, 15554, 15579, 15604, 15629, 15654, 15680, 15705, 15730, 15755, 15780, 15805, 15830, 15855, 15880, 15906, 15931, 15956, 15981, 16006, 16031, 16056, 16081, 16107, 16132, 16157, 16182, 16207, 16232, 16257, 16282, 16308, 16333, 16358, 16383};
-	int32_t * tableRead;
-
-	int32_t freq = 0;
-	uint32_t phase = 0;
-	int32_t pm = 0;
-	int32_t lastPM = 0;
-
-	inline int32_t evaluateFromFlash(uint32_t phase) {
-		int32_t index = phase >> 20;
-		int32_t frac = (phase >> 4) & 0xFFFF;
-		return fast_15_16_lerp(big_sine[index], big_sine[index + 1], frac);
-	}
-
-	inline void updatePM(int32_t pmInput) {
-		pm = pmInput - lastPM;
-		lastPM = pmInput;
-	}
-
-	inline int32_t evaluatePM(void) {
-		phase += freq + pm;
-		int32_t index = phase >> 20;
-		int32_t frac = (phase >> 4) & 0xFFFF;
-		return fast_15_16_lerp_prediff(tableRead[index], frac);
-	}
-
-	inline int32_t evaluatePMFM(int32_t fm) {
-		phase += fix16_mul(freq, fm) + pm;
-		int32_t index = phase >> 20;
-		int32_t frac = (phase >> 4) & 0xFFFF;
-		return fast_15_16_lerp_prediff(tableRead[index], frac);
-	}
-
-	inline int32_t evaluateSineBeat(int32_t newIndex) {
-		phase = newIndex * freq;
-		// scale to the wavetable size in 16 bit fixed point
-		// wavetable size is 4096 (12 bits)
-		int32_t index = phase >> 20;
-		int32_t frac = (phase >> 4) & 0xFFFF;
-		// linear interpolate between 15 bit values and return 12 bit value
-		return fast_15_16_lerp_prediff(tableRead[index], frac) >> 7;
-	}
-
-
-};
-
-
-class SyncWavetable {
-
-	int32_t previousPhase = 0;
-	int32_t previousPhaseMod = 0;
-
-public:
-
-	// assigned per mode
-	int16_t * fm;
-	int16_t * pm;
-	int16_t * pwm;
-	int16_t * morphMod;
-	int32_t cv2Offset = 0;
-	int32_t cv3Offset = 0;
-	uint32_t tableSize = 0;
-
-	// generated externally
-	int32_t phaseReset = 1;
-	int32_t increment = 0;
-	int32_t morphBase = 0;
-
-	// results
-	int32_t phaseMod = 0;
-	uint32_t phase = 0;
-	int32_t ghostPhase = 0;
-	int32_t phaseEvent = 0;
-	int32_t delta = 0;
-
-	int32_t phaseOut[32];
-	int32_t signalOut[32];
-
-	int32_t oversamplingFactor = 3;
-	int32_t bufferSize = 8;
-
-	void parseControls(ViaControls * controls);
-
-	inline int32_t incrementPhase(uint32_t * phaseDistTable);
-
-	void oversample(uint32_t * wavetable,
-			uint32_t * phaseDistTable);
-
-	void spline(uint32_t * wavetable,
-			uint32_t * phaseDistTable);
-
-	void advance(uint32_t * wavetable,
-			uint32_t * phaseDistTable) {
-		if (increment > (1 << 22)) {
-			oversample(wavetable, phaseDistTable);
-		} else {
-			spline(wavetable, phaseDistTable);
-		}
-	}
-
-
-};
-
-
-
-
-
-// simplest wavetable, provide a phase and a morph
-
-class MetaWavetable {
-
-public:
-
-	int32_t morphBase = 0;
-	int16_t * morphMod;
-	int16_t * morphScale;
-	int32_t phase = 0;
-	uint32_t tableSize = 0;
-	int32_t increment = 0;
-
-	int32_t morphModOffset = 0;
-
-	int32_t oversamplingFactor = 3;
-	int32_t bufferSize = 8;
-
-	// results
-	int32_t delta = 0;
-
-	uint32_t phaseOut[32];
-	int32_t signalOut[32];
-
-	void parseControls(ViaControls * controls);
-
-	void advance(uint32_t * wavetable) {
-		if (oversamplingFactor) {
-			advanceOversampled(wavetable);
-		} else {
-			advanceSingleSample(wavetable);
-		}
-	}
-
-	void advanceSingleSample(uint32_t * wavetable);
-
-	void advanceOversampled(uint32_t * wavetable);
-
-};
-
-
-// cheap version of that with bilinear interpolation
-
-class CheapWavetable {
-
-public:
-
-	int32_t morphBase = 0;
-	int16_t * morphMod;
-	int32_t phase = 0;
-	uint32_t tableSize = 0;
-
-	void parseControls(ViaControls * controls);
-	int32_t advance(uint32_t * wavetable);
-
-};
-
-
-/*
- *
- * Shared resources
- *
- */
-
-// pll with multiplier
-
-class PllController {
-
-	uint32_t pllCounter = 0;
-	int32_t lastMultiplier = 1;
-	int32_t lastYIndex = 0;
-
-#ifdef BUILD_F373
-	int32_t lastRatioX = 1;
-#endif
-	int32_t ratioXTransitionPoint = 0;
-	int32_t ratioXStable = 1;
-
-	int32_t ratioXHysterisis(int32_t thisRatioX, int32_t control) {
-
-		if (ratioXStable) {
-			ratioXStable = ((lastRatioX - thisRatioX) == 0);
-			ratioXTransitionPoint = control & 0b111111100000;
-			lastRatioX = thisRatioX;
-			return thisRatioX;
-		} else {
-			ratioXStable = abs(control - ratioXTransitionPoint) > 8;
-			lastRatioX = ratioXStable ? thisRatioX : lastRatioX;
-			return lastRatioX;
-		}
-
-	}
-
-#ifdef BUILD_F373
-	int32_t lastRatioY = 1;
-#endif
-	int32_t ratioYTransitionPoint = 0;
-	int32_t ratioYStable = 1;
-
-	int32_t ratioYHysterisis(int32_t control, int32_t shiftAmount) {
-
-		int32_t thisRatioY = control >> shiftAmount;
-
-		if (ratioYStable) {
-			ratioYStable = ((lastRatioY - thisRatioY) == 0);
-			ratioYTransitionPoint = thisRatioY << shiftAmount;
-			lastRatioY = thisRatioY;
-			return thisRatioY;
-		} else {
-			ratioYStable = abs(control - ratioYTransitionPoint) > 8;
-			lastRatioY = ratioYStable ? thisRatioY : lastRatioY;
-			return lastRatioY;
-		}
-
-	}
-
-public:
-
-#ifdef BUILD_VIRTUAL
-	int32_t lastRatioX = 1;
-	int32_t lastRatioY = 1;
-#endif
-
-	uint32_t virtualTimer;
-
-	uint32_t periodCount = 48000;
-	uint32_t aggregatePeriod = 48000;
-	uint32_t pileUp = 0;
-	uint32_t skipPll = 0;
-	int32_t pllNudge = 0;
-	buffer nudgeBuffer;
-	int32_t nudgeSum = 0;
-
-	uint32_t phaseSignal = 0;
-	uint32_t phaseModSignal = 0;
-	uint32_t tapTempo = 0;
-	uint32_t pllReset = 0;
-
-	int16_t * rootMod;
-	uint32_t phaseOffset = 0;
-	uint32_t syncMode = 0;
-	Scale * scale;
-	int32_t cv2Offset;
-	int32_t cv1Offset;
-
-	uint32_t fracMultiplier = 0;
-	uint32_t intMultiplier = 0;
-	uint32_t gcd = 1;
-
-	uint32_t increment = 0;
-	uint32_t phaseReset = 0;
-	uint32_t ratioChange = 0;
-	uint32_t yIndexChange = 0;
-
-	uint32_t errorSig = 0;
-
-	void parseControls(ViaControls * controls, ViaInputStreams * input);
-
-	inline void measureFrequency(void) {
-
-#ifdef BUILD_F373
-
-			// store the length of the last period
-			periodCount = TIM2->CNT;
-
-			// reset the timer value
-			TIM2->CNT = 0;
-
-#endif
-
-#ifdef BUILD_VIRTUAL
-
-		periodCount = virtualTimer;
-		virtualTimer = 0;
-
-#endif
-
-	}
-
-	void doPLL(void);
-	void generateFrequency(void);
-
-};
-
-// meta oscillator controller
-
-class MetaController {
-
-	ExpoConverter expo;
-
-public:
-
-	int32_t timeBase1 = 0;
-	int32_t timeBase2 = 0;
-	int32_t dutyCycleBase = 0;
-	int32_t triggerSignal = 0;
-	int32_t gateSignal = 0;
-	int32_t freeze = 0;
-	int32_t gateOn = 0;
-	uint32_t loopMode = 0;
-	int32_t atB = 0;
-
-	int32_t increment1 = 0;
-	int32_t increment2 = 0;
-	int32_t incrementUsed = 0;
-	int32_t dutyCycle = 0;
-	int32_t lastPhase = 0;
-	int32_t oscillatorOn = 0;
-	int16_t * fm;
-	int32_t * expoFM;
-
-	int32_t cv1Offset = 0;
-	int32_t cv2Offset = 0;
-
-	// hack
-	int32_t audioBaseIncrementStore = 34894;
-	int32_t drumBaseIncrementStore = 58623;
-	int32_t audioBaseIncrement = 34894;
-	int32_t drumBaseIncrement = 58623;
-
-	int32_t phase = 0;
-	int32_t phaseBeforeIncrement = 0;
-	int32_t ghostPhase = 0;
-	int32_t phaseEvent = 0;
-
-	void parseControlsExternal(ViaControls * controls, ViaInputStreams * inputs);
-
-	void (MetaController::*parseControls)(ViaControls * controls, ViaInputStreams * inputs);
-
-	void parseControlsAudio(ViaControls * controls, ViaInputStreams * inputs);
-	void parseControlsDrum(ViaControls * controls, ViaInputStreams * inputs);
-	void parseControlsEnv(ViaControls * controls, ViaInputStreams * inputs);
-	void parseControlsSeq(ViaControls * controls, ViaInputStreams * inputs);
-
-	void generateIncrementsExternal(ViaInputStreams * inputs);
-
-	void (MetaController::*generateIncrements)(ViaInputStreams * inputs);
-
-	void generateIncrementsAudio(ViaInputStreams * inputs);
-	void generateIncrementsDrum(ViaInputStreams * inputs);
-	void generateIncrementsEnv(ViaInputStreams * inputs);
-	void generateIncrementsSeq(ViaInputStreams * inputs);
-
-	void advancePhaseExternal(uint32_t * phaseDistTable);
-
-	int32_t (MetaController::*advancePhase)(uint32_t * phaseDistTable);
-	int32_t advancePhasePWM(uint32_t * phaseDistTable);
-	int32_t advancePhaseOversampled(uint32_t * phaseDistTable);
-
-
-	int32_t (MetaController::*incrementArbiter)(void);
-
-	int32_t noRetrigAttackState(void);
-	int32_t noRetrigReleaseState(void);
-
-	int32_t hardSyncAttackState(void);
-	int32_t hardSyncReleaseState(void);
-
-	int32_t envAttackState(void);
-	int32_t envReleaseState(void);
-	int32_t envRetriggerState(void);
-
-	int32_t gateAttackState(void);
-	int32_t gateReleaseReverseState(void);
-	int32_t gatedState(void);
-	int32_t gateReleaseState(void);
-	int32_t gateRetriggerState(void);
-
-	int32_t pendulumRestingState(void);
-	int32_t pendulumForwardAttackState(void);
-	int32_t pendulumForwardReleaseState(void);
-	int32_t pendulumReverseAttackState(void);
-	int32_t pendulumReverseReleaseState(void);
-
-	int32_t stickyPendulumRestingState(void);
-	int32_t stickyPendulumAtBState(void);
-	int32_t stickyPendulumForwardAttackState(void);
-	int32_t stickyPendulumForwardReleaseState(void);
-	int32_t stickyPendulumReverseAttackState(void);
-	int32_t stickyPendulumReverseReleaseState(void);
-
-	void (MetaController::*loopHandler)(void);
-
-	void handleLoopOff(void);
-	void handleLoopOn(void);
-
-};
-
-
-// just the envelope
-
-class SimpleEnvelope {
-
-	int32_t previousPhase;
-
-	ExpoConverter expo;
-
-public:
-
-	uint32_t attack = 100000;
-	uint32_t release = 0;
-	uint32_t increment = 0;
-	uint32_t morph = 0;
-	uint32_t phase = 0;
-	int32_t phaseEvent = 0;
-	uint32_t trigger;
-
-	int32_t * output;
-
-	void parseControls (ViaControls * controls, ViaInputStreams * inputs);
-	void advance (ViaInputStreams * inputs, uint32_t * wavetable);
-
-	int32_t (SimpleEnvelope::*incrementArbiter)(void);
-
-	int32_t attackState(void);
-	int32_t releaseState(void);
-	int32_t retriggerState(void);
-	int32_t restingState(void);
-
-};
-
-#endif /* INC_OSCILLATORS_HPP_ */
